Move chat replacement-related things into Messenger
